feat(ios): Call `onOpenWindow` for links with `target="_blank"`
On iOS, default behavior is to navigate the WebView to the target URL
when a link with `target="_blank"` is clicked

In some scenario we want to be able to intercept this from react-native
and to handle the link by ourselves (i.e. to open it into an InApp
Browser)

The `onOpenWindow` event is now called in this scenario. Then the
default "open browser" scenario does not occur